+++ src/Quartz.HttpClient/IsExternalInit.cs
@@ -1,0 +1,5 @@
+namespace System.Runtime.CompilerServices;
+
+internal static class IsExternalInit
+{
+}